# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cert Centre LMS API",
    "version": "1.0.0",
    "description": "API documentation for the Cert Centre Learning Management System"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authentication using Clerk"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "isPublished": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Chapter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "videoUrl": {
            "type": "string",
            "format": "uri"
          },
          "position": {
            "type": "integer"
          },
          "isPublished": {
            "type": "boolean"
          },
          "isFree": {
            "type": "boolean"
          },
          "courseId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "clerkAuth": []
    }
  ],
  "paths": {
...
    "/api/courses/{courseId}/chapters": {
      "post": {
        "summary": "Create a new chapter for a course",
        "tags": [
          "Chapters"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the course to add a chapter to"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The title of the chapter"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chapter created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chapter"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not logged in or not the course owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerkAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get all published chapters for a course",
        "tags": [
          "Chapters"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the course to get chapters from"
          }
        ],
        "responses": {
          "200": {
            "description": "List of published chapters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Chapter"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerkAuth": []
          }
        ]
      }
    },
    "/api/courses/{courseId}": {
      "delete": {
        "summary": "Delete a course and its associated resources",
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the course to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not logged in or not the course owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerkAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Update a course",
        "tags": [
          "Courses"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the course to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The title of the course"
                  },
                  "description": {
                    "type": "string",
                    "description": "The description of the course"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "URL for the course image"
                  },
                  "price": {
                    "type": "number",
                    "description": "Course price"
                  },
                  "categoryId": {
                    "type": "string",
                    "description": "ID of the course category"
                  },
                  "isPublished": {
                    "type": "boolean",
                    "description": "Whether the course is published"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not logged in or not the course owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerkAuth": []
          }
        ]
      }
    },
    "/api/courses": {
      "post": {
        "summary": "Create a new course",
        "tags": [
          "Courses"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The title of the course"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not logged in or not a teacher",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerkAuth": []
          }
        ]
      }
    }
  },
...
}